// packages/protocol/typescript/generated/agent_start/runtime/v1/updater_pb.d.ts
export declare type UpdaterSnapshot = Message<"agent_start.runtime.v1.UpdaterSnapshot"> & {
  /**
   * @generated from field: string app_version = 1;
   */
  appVersion: string;

  /**
   * @generated from field: string runtime_id = 2;
   */
  runtimeId: string;

  /**
   * @generated from field: agent_start.runtime.v1.UpdaterSupport support = 3;
   */
  support?: UpdaterSupport | undefined;

  /**
   * @generated from field: agent_start.runtime.v1.UpdaterStatus status = 4;
   */
  status?: UpdaterStatus | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.UpdaterSnapshot.
 * Use `create(UpdaterSnapshotSchema)` to create a new message.
 */
export declare const UpdaterSnapshotSchema: GenMessage<UpdaterSnapshot>;

/**
 * @generated from message agent_start.runtime.v1.UpdaterSupport
 */
export declare type UpdaterSupport = Message<"agent_start.runtime.v1.UpdaterSupport"> & {
  /**
   * @generated from field: agent_start.runtime.v1.UpdaterInstallMode install_mode = 1;
   */
  installMode: UpdaterInstallMode;

  /**
   * @generated from field: bool automatic = 2;
   */
  automatic: boolean;

  /**
   * @generated from field: agent_start.runtime.v1.UpdaterSupportReason reason = 3;
   */
  reason: UpdaterSupportReason;
};

/**
 * Describes the message agent_start.runtime.v1.UpdaterSupport.
 * Use `create(UpdaterSupportSchema)` to create a new message.
 */
export declare const UpdaterSupportSchema: GenMessage<UpdaterSupport>;

/**
 * @generated from message agent_start.runtime.v1.UpdaterStatus
 */
export declare type UpdaterStatus = Message<"agent_start.runtime.v1.UpdaterStatus"> & {
  /**
   * @generated from oneof agent_start.runtime.v1.UpdaterStatus.state
   */
  state: {
    /**
     * @generated from field: agent_start.runtime.v1.UpdaterIdle idle = 1;
     */
    value: UpdaterIdle;
    case: "idle";
  } | {
    /**
     * @generated from field: agent_start.runtime.v1.UpdaterChecking checking = 2;
     */
    value: UpdaterChecking;
    case: "checking";
  } | {
    /**
     * @generated from field: agent_start.runtime.v1.UpdaterAvailable available = 3;
     */
    value: UpdaterAvailable;
    case: "available";
  } | {
    /**
     * @generated from field: agent_start.runtime.v1.UpdaterNotAvailable not_available = 4;
     */
    value: UpdaterNotAvailable;
    case: "notAvailable";
  } | {
    /**
     * @generated from field: agent_start.runtime.v1.UpdaterDownloading downloading = 5;
     */
    value: UpdaterDownloading;
    case: "downloading";
  } | {
    /**
     * @generated from field: agent_start.runtime.v1.UpdaterDownloaded downloaded = 6;
     */
    value: UpdaterDownloaded;
    case: "downloaded";
  } | {
    /**
     * @generated from field: agent_start.runtime.v1.UpdaterError error = 7;
     */
    value: UpdaterError;
    case: "error";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message agent_start.runtime.v1.UpdaterStatus.
 * Use `create(UpdaterStatusSchema)` to create a new message.
 */
export declare const UpdaterStatusSchema: GenMessage<UpdaterStatus>;

/**
 * @generated from message agent_start.runtime.v1.UpdaterIdle
 */
export declare type UpdaterIdle = Message<"agent_start.runtime.v1.UpdaterIdle"> & {
};

/**
 * Describes the message agent_start.runtime.v1.UpdaterIdle.
 * Use `create(UpdaterIdleSchema)` to create a new message.
 */
export declare const UpdaterIdleSchema: GenMessage<UpdaterIdle>;

/**
 * @generated from message agent_start.runtime.v1.UpdaterChecking
 */
export declare type UpdaterChecking = Message<"agent_start.runtime.v1.UpdaterChecking"> & {
  /**
   * @generated from field: optional bool user_initiated = 1;
   */
  userInitiated?: boolean | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.UpdaterChecking.
 * Use `create(UpdaterCheckingSchema)` to create a new message.
 */
export declare const UpdaterCheckingSchema: GenMessage<UpdaterChecking>;

/**
 * @generated from message agent_start.runtime.v1.UpdaterAvailable
 */
export declare type UpdaterAvailable = Message<"agent_start.runtime.v1.UpdaterAvailable"> & {
  /**
   * @generated from field: string version = 1;
   */
  version: string;

  /**
   * @generated from field: optional string active_nudge_id = 2;
   */
  activeNudgeId?: string | undefined;

  /**
   * @generated from field: optional string release_url = 3;
   */
  releaseUrl?: string | undefined;

  /**
   * @generated from field: optional agent_start.runtime.v1.UpdaterChangelog changelog = 4;
   */
  changelog?: UpdaterChangelog | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.UpdaterAvailable.
 * Use `create(UpdaterAvailableSchema)` to create a new message.
 */
export declare const UpdaterAvailableSchema: GenMessage<UpdaterAvailable>;

/**
 * @generated from message agent_start.runtime.v1.UpdaterNotAvailable
 */
export declare type UpdaterNotAvailable = Message<"agent_start.runtime.v1.UpdaterNotAvailable"> & {
  /**
   * @generated from field: optional bool user_initiated = 1;
   */
  userInitiated?: boolean | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.UpdaterNotAvailable.
 * Use `create(UpdaterNotAvailableSchema)` to create a new message.
 */
export declare const UpdaterNotAvailableSchema: GenMessage<UpdaterNotAvailable>;

/**
 * @generated from message agent_start.runtime.v1.UpdaterDownloading
 */
export declare type UpdaterDownloading = Message<"agent_start.runtime.v1.UpdaterDownloading"> & {
  /**
   * @generated from field: uint32 percent = 1;
   */
  percent: number;

  /**
   * @generated from field: string version = 2;
   */
  version: string;

  /**
   * @generated from field: optional string active_nudge_id = 3;
   */
  activeNudgeId?: string | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.UpdaterDownloading.
 * Use `create(UpdaterDownloadingSchema)` to create a new message.
 */
export declare const UpdaterDownloadingSchema: GenMessage<UpdaterDownloading>;

/**
 * @generated from message agent_start.runtime.v1.UpdaterDownloaded
 */
export declare type UpdaterDownloaded = Message<"agent_start.runtime.v1.UpdaterDownloaded"> & {
  /**
   * @generated from field: string version = 1;
   */
  version: string;

  /**
   * @generated from field: optional string release_url = 2;
   */
  releaseUrl?: string | undefined;

  /**
   * @generated from field: optional string active_nudge_id = 3;
   */
  activeNudgeId?: string | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.UpdaterDownloaded.
 * Use `create(UpdaterDownloadedSchema)` to create a new message.
 */
export declare const UpdaterDownloadedSchema: GenMessage<UpdaterDownloaded>;

/**
 * @generated from message agent_start.runtime.v1.UpdaterError
 */
export declare type UpdaterError = Message<"agent_start.runtime.v1.UpdaterError"> & {
  /**
   * @generated from field: string message = 1;
   */
  message: string;

  /**
   * @generated from field: optional bool user_initiated = 2;
   */
  userInitiated?: boolean | undefined;

  /**
   * @generated from field: optional string active_nudge_id = 3;
   */
  activeNudgeId?: string | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.UpdaterError.
 * Use `create(UpdaterErrorSchema)` to create a new message.
 */
export declare const UpdaterErrorSchema: GenMessage<UpdaterError>;

/**
 * @generated from message agent_start.runtime.v1.UpdaterChangelog
 */
export declare type UpdaterChangelog = Message<"agent_start.runtime.v1.UpdaterChangelog"> & {
  /**
   * @generated from field: agent_start.runtime.v1.UpdaterChangelogRelease release = 1;
   */
  release?: UpdaterChangelogRelease | undefined;

  /**
   * @generated from field: optional int64 releases_behind = 2;
   */
  releasesBehind?: bigint | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.UpdaterChangelog.
 * Use `create(UpdaterChangelogSchema)` to create a new message.
 */
export declare const UpdaterChangelogSchema: GenMessage<UpdaterChangelog>;

/**
 * @generated from message agent_start.runtime.v1.UpdaterChangelogRelease
 */
export declare type UpdaterChangelogRelease = Message<"agent_start.runtime.v1.UpdaterChangelogRelease"> & {
  /**
   * @generated from field: string title = 1;
   */
  title: string;

  /**
   * @generated from field: string description = 2;
   */
  description: string;

  /**
   * @generated from field: optional string media_url = 3;
   */
  mediaUrl?: string | undefined;

  /**
   * @generated from field: string release_notes_url = 4;
   */
  releaseNotesUrl: string;
};

/**
 * Describes the message agent_start.runtime.v1.UpdaterChangelogRelease.
 * Use `create(UpdaterChangelogReleaseSchema)` to create a new message.
 */
export declare const UpdaterChangelogReleaseSchema: GenMessage<UpdaterChangelogRelease>;

/**
 * @generated from enum agent_start.runtime.v1.UpdaterInstallMode
 */
export enum UpdaterInstallMode {
  /**
   * @generated from enum value: UPDATER_INSTALL_MODE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: UPDATER_INSTALL_MODE_INTERACTIVE = 1;
   */
  INTERACTIVE = 1,

  /**
   * @generated from enum value: UPDATER_INSTALL_MODE_SUPERVISED_HEADLESS_SERVE = 2;
   */
  SUPERVISED_HEADLESS_SERVE = 2,

  /**
   * @generated from enum value: UPDATER_INSTALL_MODE_UNSUPPORTED_HEADLESS_SERVE = 3;
   */
  UNSUPPORTED_HEADLESS_SERVE = 3,
}

/**
 * Describes the enum agent_start.runtime.v1.UpdaterInstallMode.
 */
export declare const UpdaterInstallModeSchema: GenEnum<UpdaterInstallMode>;

/**
 * @generated from enum agent_start.runtime.v1.UpdaterSupportReason
 */
export enum UpdaterSupportReason {
  /**
   * @generated from enum value: UPDATER_SUPPORT_REASON_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: UPDATER_SUPPORT_REASON_AVAILABLE = 1;
   */
  AVAILABLE = 1,

  /**
   * @generated from enum value: UPDATER_SUPPORT_REASON_MANUAL_SERVICE_UPDATE_REQUIRED = 2;
   */
  MANUAL_SERVICE_UPDATE_REQUIRED = 2,

  /**
   * @generated from enum value: UPDATER_SUPPORT_REASON_UNPACKAGED_BUILD = 3;
   */
  UNPACKAGED_BUILD = 3,

  /**
   * @generated from enum value: UPDATER_SUPPORT_REASON_UPDATER_UNAVAILABLE = 4;
   */
  UPDATER_UNAVAILABLE = 4,
}